fix: fix bugs

+++ src/order/order.service.ts
@@ -1,6 +1,7 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 
 import { OrderRepository } from './repositories/order.repository';
+import { CustomerRepository } from './repositories/customer.repository';
 import { CreateOrderDto, UpdateOrderDto } from './dto/order.dto';
 import ResponseCreateOrder from './dto/responseCreateOrder.interface';
 import { PaginatedResult } from './dto/paginateResult.interface';
@@ -12,11 +13,17 @@ import { ExternalValidationService } from '../external/validation/external-valid
 export class OrderService {
   constructor(
     private readonly orderRepository: OrderRepository,
+    private readonly customerRepository: CustomerRepository,
     private readonly externalValidationService: ExternalValidationService,
   ) {}
   async getOrderById(id: number): Promise<Order | null> {
     try {
-      return await this.orderRepository.findOne(id);
+      const order = await this.orderRepository.findOne(id);
+      if (!order) {
+        // error not found
+        throw new NotFoundException('Order not found');
+      }
+      return order;
     } catch (error) {
       throw error;
     }
@@ -37,6 +44,12 @@ export class OrderService {
     createOrderDto: CreateOrderDto,
   ): Promise<ResponseCreateOrder> {
     try {
+      const customer = await this.customerRepository.getOne(
+        createOrderDto.customerId,
+      );
+      if (!customer) {
+        throw new NotFoundException('Customer not found');
+      }
       const order = await this.orderRepository.create(createOrderDto);
       const validationStatus =
         await this.externalValidationService.validateOrder();
@@ -56,6 +69,10 @@ export class OrderService {
     updateOrderDto: Partial<UpdateOrderDto>,
   ): Promise<Order> {
     try {
+      const order = await this.getOrderById(id);
+      if (!order) {
+        throw new NotFoundException('Order not found');
+      }
       return await this.orderRepository.update(id, updateOrderDto);
     } catch (error) {
       throw error;
@@ -64,6 +81,10 @@ export class OrderService {
 
   async delete(id: number): Promise<ResponseDeleteOrder> {
     try {
+      const order = await this.getOrderById(id);
+      if (!order) {
+        throw new NotFoundException('Order not found');
+      }
       await this.orderRepository.delete(id);
       return { message: 'Order deleted successfully' };
     } catch (error) {
